refactor(park-keeper): 抽離 NavOverlay 為獨立元件
- 自 Home.tsx 純搬移 NavOverlay 全螢幕羅盤導航層至 components/NavOverlay.tsx
- DirectionIcon 僅 NavOverlay 使用，一併搬移；MiniMap lazy 載入隨之遷移
- props 契約凍結：record/theme/onClose/cacheDurationDays/onPhotoOffsetChange?
- cacheDurationDays 依 S0 介面契約改為必填（唯一呼叫點原本即恆傳值，行為零變更）

測試：pnpm --filter @app/park-keeper typecheck 與 test --run（31 files / 335 tests）全綠
